Moved Pairing
Moved pairing related functions to a folder, to clean up the src folder and to organize things a bit, also updated file locations in the platformio.ini file
