+++ platformio.ini
@@ -31,10 +31,10 @@ build_src_filter =
 	+<Robot/MotorControl.h>
     +<Robot/MotorControl.cpp>
     +<Robot/Lights.h>
-    +<pairing.h>
-	+<pairing.cpp>
-    +<builtInLED.h>
-    +<builtInLED.cpp>	
+    +<Pairing/pairing.h>
+	+<Pairing/pairing.cpp>
+    +<Pairing/builtInLED.h>
+    +<Pairing/builtInLED.cpp>	
     ; -<depairingStation.cpp>
     ; -<Drive/DriveQuick.cpp>
     ; -<Drive/DriveQuick.h>
@@ -104,14 +104,14 @@ build_src_filter =
 build_src_filter = 
     -<*>
     +<depairingStation.cpp>
-    +<pairing.h>
-    +<pairing.cpp>
+    +<Pairing/pairing.h>
+    +<Pairing/pairing.cpp>
     +<PolarRobotics.h>
-    +<builtInLED.h>
-    +<builtInLED.cpp>
+    +<Pairing/builtInLED.h>
+    +<Pairing/builtInLED.cpp>
 
 [env:address]
 build_src_filter = 
 	-<*>
-	+<getAddress.cpp>
+	+<Pairing/getAddress.cpp>
